cookbook/pocketflow-tool-pdf-vision/requirements.txt
pocketflow>=0.1.0
openai>=1.0.0
PyMuPDF>=1.22.0  # 用于PDF处理
Pillow>=10.0.0   # 用于图像处理
